Break out data cleaning functions into new file.
